requirements.txt
fastapi
uvicorn
openai
pinecone
supabase==2.10.0
python-dotenv
pydantic
